requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Deep learning (for LSTM/TCN)
tensorflow>=2.13.0
keras>=2.13.0

# Satellite data
earthengine-api>=0.1.370
openeo>=0.24.0
rasterio>=1.3.0
geopandas>=0.13.0
shapely>=2.0.0

# Weather data
requests>=2.31.0
cdsapi>=0.6.0  # For ERA5 (required for ERA5 data)
xarray>=2023.1.0  # For reading ERA5 NetCDF files
netcdf4>=1.6.0  # NetCDF backend for xarray

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
folium>=0.14.0
plotly>=5.14.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
pyyaml>=6.0

